refactor: move calculations to helper
